Migra rutas de listaEspera (demanda insatisfecha) a ResourceBase
