+++ modules/turnos/index.ts
@@ -1,1 +1,2 @@
 export { InstitucionRouter } from './institucion.routes';
+export { ListaEsperaRouter } from './listaEspera.routes';
